{{cookiecutter.project_slug}}/sources/requirements/development.txt
-r base.txt

# Debug
# ---- ---- ---- ----
django-debug-toolbar==2.2
